rewrite how configs work
They now have options to create new instances of configs with defaults
or not. all fields are now Options. This enables less restrictive
configuration of configs.

Also added a new MaterializedConfig struct which is a config without any
options. This config is guaranteed to be well-formed.

Client now stores the materialized config and client::new returns an
optoin as well